+++ src/BigUInt.h
@@ -14,7 +14,7 @@ public:
 	BigUInt& operator+=(const BigUInt& a_uint);
 	BigUInt& operator*=(uint32_t a_uint);
 
-	uint32_t GetNumQuads() const { return m_quads.size(); }
+	uint32_t GetNumQuads() const { return (uint32_t)m_quads.size(); }
 	uint32_t GetQuad(uint32_t a_index) const { return m_quads[a_index]; }
 
 private:
